Adding retry for AbuseException scenario
